requirements.txt
asgiref==3.2.3
backcall==0.1.0
cffi==1.14.0
cryptography==2.8
decorator==4.4.1
Django==3.0.3
django-cors-headers==3.2.1
django-extensions==2.2.8
django-safedelete==0.5.4
django-tenant-schemas==1.10.0
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
jwt==0.6.1
mccabe==0.6.1
ordered-set==3.1.1
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.3
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.5.2
PyJWT==1.7.1
python-dotenv==0.11.0
pytz==2019.3
redis==3.4.1
six==1.14.0
sqlparse==0.3.0
traitlets==4.3.3
wcwidth==0.1.8
zappa==0.45.1
